vcenter_vm_scenario1: split playbook.yaml
The two steps a in two different isolated playbooks for the
Downstream-EE testing. The prepare.yaml depends on community.vmware
which is not part of Downstream-EE.
This allow use to run the first stage with a regular ansible, and
the second part from within a Downstream-EE container.
Please enter the commit message for your changes. Lines starting