Change URL for library item list in 7.0.3 spec (#464)
Change URL for library item list in 7.0.3 spec

SUMMARY
When retrieving a list of items inside a content library, the URL was malformed due to a double key in the query. Spec 7.0.3 is changed to look like it did in 7.0.2.
Fixes #463
ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME
content_library_item_info

Reviewed-by: GomathiselviS
Reviewed-by: Alina Buzachis