Add eco vcenter upstream ci capabilities (#538)
SUMMARY
This PR aims to insert self hosted runner to enable integration tests to run on the eco vcenter environment.
This includes several adjustments to the code and additional action to be triggered for each PR, and run all the
integration tests on that env with the PR's new code.

Reviewed-by: mikemorency
Reviewed-by: Anna Savina