Enhance appliance health target to run on vCenter env (#545)
https://issues.redhat.com/browse/ACA-1884

This PR updates the appliance_health target to support execution in a vCenter environment.

Since the appliance and appliance-multi-hosts targets seems to be disabled in Zuul (at this time), the intention for these target changes is to eventually replace the original targets. Once this transition is complete, they will be removed from Zuul, and only the eco-vcenter-ci will be supported. Please share your thoughts and opinions on this approach.

Reviewed-by: Anna Savina
Reviewed-by: Polina Rabinovich
Reviewed-by: Shelly Miron
Reviewed-by: mikemorency