fix nonsensical path results in lookup plugin (#550)
SUMMARY
If someone specifies a path to an object (like a cluster) that cannot contain the lookup object, the plugin ignores the cluster name and just uses the datacenter as a filter.
Ideally, I think we would throw an error telling the user that they are searching for a path the doesnt make sense. But the old plugin output is just an empty string, so I am keeping that output.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
lookup plugins

Reviewed-by: Anna Savina
Reviewed-by: Danielle Barda
Reviewed-by: mikemorency