disable appliance_networking test due to the bug (#575)
This PR temporarily disables the appliance_networking test that has recently started failing in Shelly's PR. The failure appears to be due to an environmental change, as the test was passing previously without issues.
The failure also revealed a bug in the vmware.vmware_rest.appliance_networking module. Specifically, when attempting to disable IPv6, the error message is misleading, stating:
"Failed to enable ipv6 for network interfaces."
A bug has been opened to track this issue - link

Reviewed-by: Shelly Miron